Fix "frame of death" bug, ie status request frame on monitor.
In some cases it may happen that we receive a status request frame up
from the monitor socket.
The code doesn't handle this case well (read: it will crash), so we
ignore this kind of frames.

All tests pass and code scrutiny doesn't reveal any device that
currently use this functionality.
There were also some devices that already did this check in case
there were some other touchscreens attached on the same bus, so it seems
safe to do it for every device in one place.

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>